error handling and remove taxonomyId request parameter in content bundle api

+++ platform-modules/taxonomy/module/taxonomy-mgr/src/main/java/com/ilimi/taxonomy/util/HttpDownloadUtility.java
@@ -10,6 +10,7 @@ import java.util.List;
 
 /**
  * A utility that downloads a file from a URL.
+ * 
  * @author Mohammad Azharuddin
  *
  */
@@ -18,12 +19,14 @@ public class HttpDownloadUtility {
 
     /**
      * Downloads a file from a URL
-     * @param fileURL HTTP URL of the file to be downloaded
-     * @param saveDir path of the directory to save the file
+     * 
+     * @param fileURL
+     *            HTTP URL of the file to be downloaded
+     * @param saveDir
+     *            path of the directory to save the file
      * @throws IOException
      */
-    public static File downloadFile(String fileURL, String saveDir)
-            throws IOException {
+    public static File downloadFile(String fileURL, String saveDir) throws IOException {
         URL url = new URL(fileURL);
         HttpURLConnection httpConn = (HttpURLConnection) url.openConnection();
         int responseCode = httpConn.getResponseCode();
@@ -32,26 +35,19 @@ public class HttpDownloadUtility {
         if (responseCode == HttpURLConnection.HTTP_OK) {
             String fileName = "";
             String disposition = httpConn.getHeaderField("Content-Disposition");
-            String contentType = httpConn.getContentType();
-            int contentLength = httpConn.getContentLength();
+            httpConn.getContentType();
+            httpConn.getContentLength();
 
             if (disposition != null) {
                 // extracts file name from header field
                 int index = disposition.indexOf("filename=");
                 if (index > 0) {
-                    fileName = disposition.substring(index + 10,
-                            disposition.length() - 1);
+                    fileName = disposition.substring(index + 10, disposition.length() - 1);
                 }
             } else {
                 // extracts file name from URL
-                fileName = fileURL.substring(fileURL.lastIndexOf("/") + 1,
-                        fileURL.length());
+                fileName = fileURL.substring(fileURL.lastIndexOf("/") + 1, fileURL.length());
             }
-
-            System.out.println("Content-Type = " + contentType);
-            System.out.println("Content-Disposition = " + disposition);
-            System.out.println("Content-Length = " + contentLength);
-            System.out.println("fileName = " + fileName);
 
             // opens input stream from the HTTP connection
             InputStream inputStream = httpConn.getInputStream();
@@ -68,8 +64,6 @@ public class HttpDownloadUtility {
 
             outputStream.close();
             inputStream.close();
-
-            System.out.println("File downloaded");
             return new File(saveFilePath);
         } else {
             System.out.println("No file to download. Server replied HTTP code: " + responseCode);
@@ -77,23 +71,23 @@ public class HttpDownloadUtility {
         httpConn.disconnect();
         return null;
     }
-    
+
     public static InputStream downloadFile(String fileURL) {
-    	try {
-	    	final InputStream is = (new URL(fileURL).openConnection()).getInputStream();
-	    	return is;
-    	} catch(IOException ioe) {
-    		return null;
-    	}
+        try {
+            final InputStream is = (new URL(fileURL).openConnection()).getInputStream();
+            return is;
+        } catch (IOException ioe) {
+            return null;
+        }
     }
-    
+
     public static void DeleteFiles(List<File> files) {
-    	for (File file : files) {
-	    	if(file.delete()){
-				System.out.println(file.getName() + " is deleted!");
-			}else{
-				System.out.println("Delete operation is failed.");
-			}
-    	}
+        for (File file : files) {
+            if (file.exists()) {
+                if (file.delete()) {
+                    System.out.println(file.getName() + " is deleted!");
+                }
+            }
+        }
     }
 }
